git-svn-id: file:///Users/MacAir/Dropbox/xolotl-cleaned/trunk@703 72105f29-0fe1-43bd-9363-b1e281e962a2

+++ xolotl/xolotlSolver/Monitor.cpp
@@ -241,7 +241,7 @@ PetscErrorCode computeHeliumFluence(TS ts, PetscInt timestep, PetscReal time,
 	checkPetscError(ierr);
 
 	// The length of the time step
-	float dt = time - previousTime;
+	double dt = time - previousTime;
 
 	// Increment the fluence with the value at this current timestep
 	fluxHandler->incrementHeFluence(dt);
